mailscout-main/requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
mailscout==0.1.1
MarkupSafe==3.0.2
Unidecode==1.3.8
Werkzeug==3.1.3
